* Install all dependencies from panoptes-utils. * Move star finding info to `sources`.
