+++ requirements.txt
@@ -3,6 +3,6 @@ astropy
 astroquery
 matplotlib
 numpy
-panoptes-utils>=0.0.7
+panoptes-utils[all]>=0.0.7
 pandas
 tqdm
